requirements.txt
prefect~=2.8.7
python-dotenv
pandas
google-cloud-storage
pyarrow~=11.0.0